mediatek: mac sgmii to 2500basex mode mt7531
For port 5 of mt7531 if connected to an PHY port (like with Netgear WAX206)
allow switching between SGMII and 2500BASE-X mode if link speed is 2500;
and back to SGMII.

This only works if in the devicetree, for port 5 phy-mode is set to sgmii.

And it only works when booting into the initramfs image. The regular
sysupdate image doesn't work (and doesn't output any printk()'s).

Signed-off-by: Thomas Kupper <thomas.kupper@gmail.com>